+++ src/main/kotlin/com/esmartit/seendevicesdatastore/domain/TotalDevices.kt
@@ -2,4 +2,4 @@ package com.esmartit.seendevicesdatastore.domain
 
 import java.time.Instant
 
-data class TotalDevices(val count: Long, val time: Instant)
+data class TotalDevices(val count: Int, val time: Instant)
